Make abstract OPC filetype, make UFP a child of that
