+++ tests/filetypes/__init__.py
@@ -1,0 +1,2 @@
+# Copyright (c) 2018 Ultimaker B.V.
+# libCharon is released under the terms of the LGPLv3 or higher.
